minor improvements el7mdullah & inshallah
