+++ .idea/gradle.xml
@@ -22,6 +22,7 @@
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/articles" />
             <option value="$PROJECT_DIR$/feature/combinedModules" />
+            <option value="$PROJECT_DIR$/feature/shared" />
             <option value="$PROJECT_DIR$/ui" />
             <option value="$PROJECT_DIR$/ui/articles" />
             <option value="$PROJECT_DIR$/ui/shared" />
