+++ gradle/libs.versions.toml
@@ -47,6 +47,7 @@ kotlinx-serialization-json = "1.8.1"
 mockk = "1.14.2"
 org-junit = "5.13.1"
 byte-buddy-agent = "1.14.17"
+material = "1.12.0"
 
 [libraries]
 
@@ -108,6 +109,7 @@ kotlinx-coroutines-test = { group = "org.jetbrains.kotlinx", name = "kotlinx-cor
 androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitVersion" }
 androidx-espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espressoCore" }
 androidx-ui-test-junit4 = { group = "androidx.compose.ui", name = "ui-test-junit4" }
+material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
